test: refactor plugin setup (#31)
I found out you can actually pass plugin instance to pytester when
running it inline: this means we can grab the list of spans in a
sub-pytest if we run it inline.

This simplifies the overall architecture to split the tests in two mode:
- either we run them in a subprocess to test the whole lifecycle (e.g.
  upload errors are printed)
- either we run them inprocess and we can read the list of spans

This will make it easier in the future to replace pytest-opentelemetry
and test the spans that are emitted.