chore: add support PYTEST_MERGIFY_ENABLED=true (#131)
CI=true is good to enable it by default in CI. But some tests may change
their behavior when CI is set.

For example in pytest, they use tox and ensure CI is not passed
to the test suite. Enabling pytest-mergify breaks the test suite.

This change introduces an alternative for user that can't rely
on CI=true.