fix: better error report when token/repo are empty (#138)
It's very easy in GitHub Actions to do:

run:
    env:
      - MERGIFY_TOKEN: ${{ secrets.MERGIFY_TOKEN }}

And having secrets.MERGIFY_TOKEN empty because of typo, secrets permission issues, misconfigured reusable workflow.

This configuration effectively set the the env but it's an empty string.

The current code don't start a tracer as expected, but didn't report anything.

This adds:
* the reporting for when repo_name and token are empty.
* an generic error message if the tracer didn't start for some reason and we didn't provide the root cause.

Fixes MRGFY-5471